Add support for AXS103 board
This is a real development board from Synopsys with much more peripherals
compared to nSIM simulator.

Among other things we have:
 - 100 Mbit Ethernet port
 - SD-card slot
 - USB 2.0 slot

With this board in OE we'll be able to test installation of external .ipks.

The board could be selected as simple as setting MACHINE="axs103" in
conf/local.conf.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>