linux: Allow usage of any ARC tools independent of its default ARCH
We used to use a separate toolchain for ARCompact and ARCv2 cores
as they typically contain libgcc.a built for the particular architecture.
Thsi is very true for a standalone toolchain but not always true for modern
build-systems.

So we fixup the kernel sources to resolve that dependency.
And with the change in place we now may finally reuse one built toolchain
for both ARCompact (read ARC700) and ARCv2 (read ARc HS38) targets.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>