openl2tp: Fix uClibc compilation issue
With uClibc, the package fails to build because of IPV6 not supported
but that are defined as default in RPC functions.
Added the patches that exist in Pyro's branch of OE-core:
http://cgit.openembedded.org/meta-openembedded/tree/meta-networking/recipes-protocols/openl2tp/openl2tp_1.8.bb?h=pyro

The CPPFLAGS and OPTFLAGS need also to be set to use the target's ones
and not the host's ones.

Thanks to that, the package builds correctly.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>