conf/machine: Make architecture DEFAULTTUNE even weaker
We have DEFAULTTUNE weakly defined in both architecture "include"
and in the board configuration. That's because arch's default is
a safety net for everythign, while board's default tunes for the board
capabilities.

But given "weakness" level "?=" is the same in both the places
board's config doesn't override arch's.

And while we may hard set DEFAULTTUNE for the particular board,
what if we want to build with say simpler options for some reason.

For that we need to have a way to override DEFAULTTUNE in local.conf.
Luckily there're 2 levels of "weakness" of variable assignments in
bitbake, see https://docs.yoctoproject.org/bitbake/bitbake-user-manual/bitbake-user-manual-metadata.html#setting-a-weak-default-value.

So we set arch's DEAFULTTUNE as the "weakest" & board's just "weak".

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>